consolidated get_dashboard_order method to rest framework widgets file.
