+++ src/sentry/api/endpoints/organization_dashboard_widgets.py
@@ -5,10 +5,10 @@ from rest_framework.response import Response
 
 from sentry.api.base import DocSection
 from sentry.api.bases.dashboard import (
-    OrganizationDashboardEndpoint, get_next_dashboard_order
+    OrganizationDashboardEndpoint
 )
 from sentry.api.serializers import serialize
-from sentry.api.serializers.rest_framework import WidgetSerializer
+from sentry.api.serializers.rest_framework import get_next_dashboard_order, WidgetSerializer
 from sentry.models import Widget, WidgetDataSource
